refactor: theme implementation
