+++ src/main/java/run/halo/app/infra/SystemConfigurableEnvironmentFetcher.java
@@ -1,18 +1,12 @@
 package run.halo.app.infra;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.lang.NonNull;
-import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
 import run.halo.app.extension.ConfigMap;
 import run.halo.app.extension.ExtensionClient;
-import run.halo.app.infra.utils.JsonParseException;
 import run.halo.app.infra.utils.JsonUtils;
 
 /**
@@ -32,20 +26,6 @@ public class SystemConfigurableEnvironmentFetcher {
         this.conversionService = conversionService;
     }
 
-    @Nullable
-    @SuppressWarnings("unchecked")
-    public <T> T get(SystemSetting systemSetting) {
-        Class<?> valueType = systemSetting.getValueType();
-        String value = getInternal(systemSetting.getGroup());
-        if (value == null) {
-            return null;
-        }
-        if (valueType.isPrimitive()) {
-            return (T) conversionService.convert(value, valueType);
-        }
-        return (T) JsonUtils.jsonToObject(value, valueType);
-    }
-
     public <T> Optional<T> fetch(String key, Class<T> type) {
         var stringValue = getInternal(key);
         if (stringValue == null) {
@@ -59,17 +39,6 @@ public class SystemConfigurableEnvironmentFetcher {
 
     private String getInternal(String group) {
         return getValuesInternal().get(group);
-    }
-
-    private JsonNode readTree(String json) {
-        if (StringUtils.isBlank(json)) {
-            return JsonNodeFactory.instance.missingNode();
-        }
-        try {
-            return JsonUtils.DEFAULT_JSON_MAPPER.readTree(json);
-        } catch (JsonProcessingException e) {
-            throw new JsonParseException(e);
-        }
     }
 
     @NonNull
