+++ src/main/java/run/halo/app/infra/SystemSetting.java
@@ -3,30 +3,17 @@ package run.halo.app.infra;
 import lombok.Data;
 
 /**
+ * TODO Optimization value acquisition.
+ *
  * @author guqing
  * @since 2.0.0
  */
-public enum SystemSetting {
-    THEME("theme", Theme.class);
-
-    private final String group;
-    private final Class<?> valueType;
-
-    SystemSetting(String group, Class<?> valueType) {
-        this.group = group;
-        this.valueType = valueType;
-    }
-
-    public String getGroup() {
-        return group;
-    }
-
-    public Class<?> getValueType() {
-        return valueType;
-    }
+public class SystemSetting {
 
     @Data
     public static class Theme {
+        public static final String GROUP = "theme";
+
         private String active;
     }
 }
