+++ src/test/java/run/halo/app/theme/ThemeContextTest.java
@@ -1,5 +1,6 @@
-package run.halo.app.theme.newplan;
+package run.halo.app.theme;
 
+import java.nio.file.Paths;
 import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
@@ -14,32 +15,16 @@ import run.halo.app.infra.utils.JsonUtils;
 class ThemeContextTest {
 
     @Test
-    void constructorTest() throws JSONException {
-        ThemeContext testTheme =
-            new ThemeContext("testTheme", "/tmp/themes/testTheme", false);
-        String s = JsonUtils.objectToJson(testTheme);
-        JSONAssert.assertEquals("""
-                {
-                    "themeName": "testTheme",
-                    "path": "file:///tmp/themes/testTheme",
-                    "active": false
-                }
-                """,
-            s,
-            false);
-    }
-
-    @Test
     void constructorBuilderTest() throws JSONException {
         ThemeContext testTheme = ThemeContext.builder()
             .name("testTheme")
-            .path("/tmp/themes/testTheme")
+            .path(Paths.get("/tmp/themes/testTheme"))
             .active(true)
             .build();
         String s = JsonUtils.objectToJson(testTheme);
         JSONAssert.assertEquals("""
                 {
-                    "themeName": "testTheme",
+                    "name": "testTheme",
                     "path": "file:///tmp/themes/testTheme",
                     "active": true
                 }
