+++ src/test/java/run/halo/app/theme/message/ThemeMessageResolutionUtilsTest.java
@@ -3,14 +3,14 @@ package run.halo.app.theme.message;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.FileNotFoundException;
-import java.net.URISyntaxException;
 import java.net.URL;
+import java.nio.file.Paths;
 import java.util.Locale;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.util.ResourceUtils;
-import run.halo.app.theme.newplan.ThemeContext;
+import run.halo.app.theme.ThemeContext;
 
 /**
  * @author guqing
@@ -25,7 +25,7 @@ class ThemeMessageResolutionUtilsTest {
     }
 
     @Test
-    void resolveMessagesForTemplateForDefault() throws URISyntaxException {
+    void resolveMessagesForTemplateForDefault() {
         Map<String, String> properties =
             ThemeMessageResolutionUtils.resolveMessagesForTemplate(Locale.CHINESE, getTheme());
         assertThat(properties).hasSize(1);
@@ -33,7 +33,7 @@ class ThemeMessageResolutionUtilsTest {
     }
 
     @Test
-    void resolveMessagesForTemplateForEnglish() throws URISyntaxException {
+    void resolveMessagesForTemplateForEnglish() {
         Map<String, String> properties =
             ThemeMessageResolutionUtils.resolveMessagesForTemplate(Locale.ENGLISH, getTheme());
         assertThat(properties).hasSize(1);
@@ -48,10 +48,10 @@ class ThemeMessageResolutionUtilsTest {
         assertThat(s).isEqualTo("Welcome Halo to the index");
     }
 
-    ThemeContext getTheme() throws URISyntaxException {
+    ThemeContext getTheme() {
         return ThemeContext.builder()
             .name("default")
-            .path(defaultThemeUrl.toString())
+            .path(Paths.get(defaultThemeUrl.getPath()))
             .active(true)
             .build();
     }
